% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft.R
\docType{package}
\name{DRAFT-package}
\alias{DRAFT}
\alias{DRAFT-package}
\title{DRAFT: Disease Rapid Analysis and Forecasting Tool}
\description{
Using an MCMC procedure, fit an SIR, SEIR or behavior modification SIR model 	to user-provided incidence data.
}
\details{
The primary function that a user will need is runbSEIR().  This function expects the user to input an incidence data.frame with columns 'date' and 'cases'.  For non-uniform reporting, the code assumes that the reporting period for the nth row runs from the date[n-1]+1 to date[n].  The 'date' column class should be "Date".
}
\author{
\strong{Maintainer}: James Turtle \email{jturtle@predsci.com}

Authors:
\itemize{
  \item Michal Ben-Nun \email{mbennun@predsci.com}
}

Other contributors:
\itemize{
  \item Pete Riley \email{pete@predsci.com} [contributor]
}

}
\keyword{internal}
