% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{incidence_data1}
\alias{incidence_data1}
\title{Sample input data.frame for DRAFT::runbSEIR()}
\format{A data.frame with columns \code{"date"} and \code{"cases"}.  The \code{"date"} column is of class "Date" as expected by \code{runbSEIR()}.}
\usage{
data(incidence_data1)
}
\description{
Incidence data for San Diego County confirmed influenza cases from the 2017-2018 season.
}
\keyword{datasets}
