% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{incidence_data2}
\alias{incidence_data2}
\title{Sample input data.frame for DRAFT::runbSEIR()}
\format{A data.frame with columns \code{"date"} and \code{"cases"}.  The \code{"date"} column is of class "Date" as expected by \code{runbSEIR()}.}
\usage{
data(incidence_data2)
}
\description{
Incidence data for San Diego County confirmed influenza cases from the 2014-2015 season.  Trailing rows of 'cases' column are NA indicating that a forecast is requested from \code{DRAFT}.
}
\keyword{datasets}
