% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{incidence_data4}
\alias{incidence_data4}
\title{Sample input data.frame for DRAFT::runbSEIR()}
\format{A data.frame with columns \code{"date"} and \code{"cases"}.  The \code{"date"} column is of class "Date" as expected by \code{runbSEIR()}.}
\usage{
data(incidence_data4)
}
\description{
Incidence data for Liberia ebola cases from the 2013 outbreak.  Trailing rows of 'cases' column are NA indicating that a forecast is requested from \code{DRAFT}.
}
\keyword{datasets}
