% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_user_data_fxns.R
\name{runDRAFT}
\alias{runDRAFT}
\title{Main Driver for using \code{DRAFT} for Fitting User Provided Incidence}
\usage{
runDRAFT(inc_data = NULL, pop = 10000, epi_model = 1, Tg = 3,
  sigma = NULL, dp = NULL, dq = NULL, ts = NULL, dL = NULL)
}
\arguments{
\item{inc_data}{Dataframe containing incidence data.  Must contain 'date' and 'cases' columns.  The 'date' column must either be Date-class or convert to Date-class using as.Date(inc_data$date, format="\%Y-\%m-\%d").}

\item{pop}{Integer population of the region for which incidence is provided}

\item{epi_model}{- integer 1 (SIR), 2 (SEIR), 3 (SIR with behavior terms). Default is 1-SIR.}

\item{Tg}{Numeric, generation time in days. Default is 3 days}

\item{sigma}{inverse of of the latent period in days. Needed only for an SEIR model. Default NULL}

\item{dp}{Proporiton of susceptible contact rate (0-1) following behavior modification. For example: If susceptibles reduce their contacts by 25\%, set \code{dp=0.75}.  Only used if inc_data requires a forecast.}

\item{dq}{Proportion of infectious contact rate (0-1) following behavior modification.  If infectious cases reduce their contacts by one half, set \code{dq=0.5}.  Only used if inc_data requires a forecast.}

\item{ts}{Date class.  Start date of behavior modification.}

\item{dL}{Number of days for behavior modification to completely take effect.}
}
\value{
results A list with the input and entire output of the run.
}
\description{
\code{runDRAFT} reads in a user provided incidence file and uses that along with the user
provided population size and value for the generation time, Tg, and if relevant the latent period
sigma to fit, and if desired, forecast the incidence.
Data cadence is arbitrary but at most can be monthly
We support only S-I-R and S-E-I-R models for a single population with a fixed force of infection
The incidence file must have two columns: dates and cases
All other parameters are set by the code
}
\examples{
# Run an SEIR model using the incidence file and assuming a population of 1 million people.
# The generation time and latent period are set to 2.6 days and 3 days respectively
head(incidence_data2)
\dontrun{output <- runDRAFT(inc_data=incidence_data2, pop = 1e6, 
epi_model = 2, Tg = 2.6, sigma = 3.)}

# Run an SIR model using the incidence file and assuming a population of 10,000 people.
# The generation time is set to 3 days. (No need to define a latent period.)
head(incidence_data1)
\dontrun{output <- runDRAFT(inc_data=incidence_data2, pop = 1e5, epi_model = 1, Tg = 3.)}

}
